Rename upload.js to Upload.js
